Time: 17 ms (43.47%), Space: 43.3 MB (74.6%) - LeetHub

+++ 0437-path-sum-iii/0437-path-sum-iii.java
@@ -25,7 +25,7 @@ class Solution {
         return ans;
     }
 
-    public void helper(TreeNode root, int sum){
+    public void helper(TreeNode root, long sum){
         if(root==null)return;
         if(sum == root.val)ans++;
